Test and create dataset and configuration files
Finished the initial testing of how NLP text classfication work and derived a initial prediction for the  dataset. Remember that the the training are not official. That is to say all of them are trained on a sub sample to make testing the pipe faster for the moment before obtaininng the finial result. Remember to change such.